requirements.txt
starlette
httpx
uvicorn
glom
loguru
pytest-asyncio
pytest
aiosqlite
python-multipart
pillow
